+++ app/src/main/java/com/triviagenai/triviagen/trivia/presentation/mainmenu/MainMenuScreen.kt
@@ -1,7 +1,6 @@
 package com.triviagenai.triviagen.trivia.presentation.mainmenu
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -12,16 +11,14 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.navigation.NavHostController
-import com.triviagenai.triviagen.MainActivity
 import com.triviagenai.triviagen.R
-import com.triviagenai.triviagen.core.presentation.Route
+import com.triviagenai.triviagen.core.presentation.navigation.Route
 import com.triviagenai.triviagen.core.presentation.TriviaGenScaffold
 
 @Composable
